cleanup(core): extract extractFileFromTarball to its own module
fileutils.ts imported tar-stream and zlib just for extractFileFromTarball.
Importers of readJsonFile / fileExists / etc. shouldn't drag in tar-stream
on every load. Move the tarball helper to utils/tar.ts and update the
only two callers (migrate, prompt-files).
