+++ packages/nx/src/utils/fileutils.ts
@@ -1,8 +1,6 @@
 import { parseJson, serializeJson } from './json';
 import type { JsonParseOptions, JsonSerializeOptions } from './json';
 import {
-  createReadStream,
-  createWriteStream,
   PathLike,
   readFileSync,
   writeFileSync,
@@ -12,8 +10,6 @@ import {
 } from 'node:fs';
 import { mkdir, writeFile } from 'node:fs/promises';
 import { dirname } from 'path';
-import * as tar from 'tar-stream';
-import { createGunzip } from 'zlib';
 
 export interface JsonReadOptions extends JsonParseOptions {
   /**
@@ -151,52 +147,6 @@ export function isRelativePath(path: string): boolean {
     path.startsWith('./') ||
     path.startsWith('../')
   );
-}
-
-/**
- * Extracts a file from a given tarball to the specified destination.
- * @param tarballPath The path to the tarball from where the file should be extracted.
- * @param file The path to the file inside the tarball.
- * @param destinationFilePath The destination file path.
- * @returns True if the file was extracted successfully, false otherwise.
- */
-export async function extractFileFromTarball(
-  tarballPath: string,
-  file: string,
-  destinationFilePath: string
-) {
-  return new Promise<string>((resolve, reject) => {
-    mkdirSync(dirname(destinationFilePath), { recursive: true });
-    var tarExtractStream = tar.extract();
-    const destinationFileStream = createWriteStream(destinationFilePath);
-
-    let isFileExtracted = false;
-    tarExtractStream.on('entry', function (header, stream, next) {
-      if (header.name === file) {
-        stream.pipe(destinationFileStream);
-        stream.on('end', () => {
-          isFileExtracted = true;
-        });
-        destinationFileStream.on('close', () => {
-          resolve(destinationFilePath);
-        });
-      }
-
-      stream.on('end', function () {
-        next();
-      });
-
-      stream.resume();
-    });
-
-    tarExtractStream.on('finish', function () {
-      if (!isFileExtracted) {
-        reject();
-      }
-    });
-
-    createReadStream(tarballPath).pipe(createGunzip()).pipe(tarExtractStream);
-  });
 }
 
 export function readFileIfExisting(path: string) {
